refactor: separate MCP preview handling
