+++ packages/cli/src/commands/mcp.ts
@@ -40,7 +40,6 @@ import {
 import { executeProjectSessionApiOperation } from "../project-session-api";
 import { createPreviewController, findAvailablePort } from "../preview-server";
 import {
-  captureScreenshotWithBrowserInstall,
   createScreenshotCaptureSession,
   installTesseractForOcr,
 } from "../screenshot";
@@ -56,7 +55,6 @@ import {
 import { readCliDoc } from "../docs";
 import { printJson } from "../json-output";
 import { isPlainRecord } from "../type-utils";
-import { inspectGeneratedBuildMetrics } from "../generated-build-metrics";
 import { getLocalProjectStateDirectory, LOCAL_DATA_FILE } from "../config";
 import {
   assertMcpBatchMutationApproved,
@@ -67,12 +65,7 @@ import {
 } from "./mcp-batch";
 import { apiCompatibilityHeaders } from "./api";
 import { importProject as importProjectCommand } from "./import";
-import {
-  preparePreviewProject,
-  previewDefaultTemplate,
-  type PreviewSource,
-  validatePreviewServerOptions,
-} from "./preview";
+import { createMcpPreviewHandlers } from "./mcp-preview";
 import type {
   CommonYargsArgv,
   StrictYargsOptionsToInterface,
@@ -82,52 +75,6 @@ type StartableProjectSession = {
   snapshot: ProjectSessionSnapshot | undefined;
   initialize: () => Promise<unknown>;
   markStale: (namespaces: readonly BuilderNamespace[]) => Promise<unknown>;
-};
-
-type McpPreviewController = Pick<
-  ReturnType<typeof createPreviewController>,
-  "startAndWait" | "status" | "resolveUrl"
-> &
-  Partial<Pick<ReturnType<typeof createPreviewController>, "stop">>;
-
-type McpPreviewInput = {
-  host?: string;
-  port?: number;
-  source?: PreviewSource;
-  imageDomains?: string[];
-};
-
-type CaptureScreenshotInput = Parameters<
-  typeof captureScreenshotWithBrowserInstall
->[0];
-
-type McpScreenshotInput = {
-  url?: string;
-  baseUrl?: string;
-  path?: string;
-  output?: string;
-  host?: string;
-  port?: number;
-  imageDomains?: string[];
-  viewport: { width: number; height: number };
-  fullPage?: boolean;
-  includeImageMetrics?: boolean;
-  includeResourceMetrics?: boolean;
-  includeContrastMetrics?: boolean;
-  browser?: CaptureScreenshotInput["browser"];
-  browserPath?: string;
-  waitUntil?: CaptureScreenshotInput["waitUntil"];
-  waitForSelector?: string;
-  waitForTimeout?: number;
-  timeout?: number;
-  source?: PreviewSource;
-  format?: "png" | "jpeg" | "webp";
-  quality?: number;
-  scale?: number;
-};
-
-type McpToolProgress = {
-  report: (message: string) => void;
 };
 
 type PublicApiCommand = (typeof publicApiOperations)[number]["command"];
@@ -344,20 +291,6 @@ const publicApiOperationByCommand = new Map<
 const shouldInvalidatePreview = (command: string) =>
   publicApiOperationByCommand.get(command)?.method === "mutation";
 
-const prepareDefaultPreviewProject = (
-  source: PreviewSource = "local",
-  prepareSessionDataFile?: () => Promise<void>
-) =>
-  preparePreviewProject({
-    assets: true,
-    template: [...previewDefaultTemplate],
-    generate: true,
-    source,
-    syncIfMissing: true,
-    silent: true,
-    prepareSessionDataFile,
-  });
-
 const getLoadedProjectSessionSnapshot = (session: StartableProjectSession) => {
   const snapshot = session.snapshot;
   if (snapshot === undefined) {
@@ -366,239 +299,6 @@ const getLoadedProjectSessionSnapshot = (session: StartableProjectSession) => {
     );
   }
   return snapshot;
-};
-
-const createMcpPreviewHandlers = ({
-  preview,
-  isStale = () => true,
-  markFresh = () => undefined,
-  preparePreview = prepareDefaultPreviewProject,
-  prepareSessionDataFile,
-  captureScreenshot = captureScreenshotWithBrowserInstall,
-  createCaptureSession,
-}: {
-  preview: McpPreviewController;
-  isStale?: () => boolean;
-  markFresh?: () => void;
-  preparePreview?: typeof prepareDefaultPreviewProject;
-  prepareSessionDataFile?: () => Promise<void>;
-  captureScreenshot?: typeof captureScreenshotWithBrowserInstall;
-  createCaptureSession?: typeof createScreenshotCaptureSession;
-}) => {
-  let captureSession:
-    | ReturnType<typeof createScreenshotCaptureSession>
-    | undefined;
-  const closeCaptureSession = async () => {
-    try {
-      await captureSession?.close();
-    } finally {
-      captureSession = undefined;
-    }
-  };
-  const rejectBuilderUrl = (value: string) => {
-    let url: URL;
-    try {
-      url = new URL(value);
-    } catch {
-      return;
-    }
-    const isProjectHost =
-      url.hostname.startsWith("p-") &&
-      (url.hostname.endsWith(".wstd.dev") ||
-        url.hostname.endsWith(".apps.webstudio.is"));
-    if (
-      isProjectHost &&
-      url.searchParams.has("authToken") &&
-      url.searchParams.has("mode")
-    ) {
-      throw Object.assign(
-        new Error(
-          "This is an authenticated Builder/share URL, not a generated-site preview. Use the share URL with webstudio init --link, then capture a generated route with screenshot { path }, or pass an intentional standalone site URL without Builder authentication parameters."
-        ),
-        { code: "BUILDER_URL_IS_NOT_SITE_PREVIEW" }
-      );
-    }
-  };
-  const resolveScreenshotUrl = async (
-    input: McpScreenshotInput,
-    progress?: McpToolProgress
-  ) => {
-    if (input.url !== undefined) {
-      rejectBuilderUrl(input.url);
-      return input.url;
-    }
-    if (input.path === undefined) {
-      throw new Error("MCP screenshot requires url or path.");
-    }
-    if (input.baseUrl !== undefined) {
-      rejectBuilderUrl(input.baseUrl);
-      return new URL(input.path, input.baseUrl).toString();
-    }
-    const hasExplicitPreviewTarget =
-      input.host !== undefined || input.port !== undefined;
-    if (
-      preview.status().running === false ||
-      isStale() ||
-      hasExplicitPreviewTarget
-    ) {
-      await closeCaptureSession();
-      validatePreviewServerOptions({
-        host: input.host ?? "127.0.0.1",
-        port: input.port ?? 5173,
-        imageDomains: input.imageDomains,
-      });
-      progress?.report("tool screenshot preparing generated preview project");
-      const previewProject = await preparePreview(
-        input.source,
-        prepareSessionDataFile
-      );
-      progress?.report("tool screenshot starting production preview server");
-      await preview.startAndWait({
-        cwd: previewProject.cwd,
-        buildCacheKey: previewProject.buildCacheKey,
-        host: input.host,
-        port: input.port,
-        imageDomains: input.imageDomains,
-        restart: true,
-      });
-      markFresh();
-    }
-    return preview.resolveUrl(input.path);
-  };
-  const getCaptureOptions = (input: McpScreenshotInput, url: string) => ({
-    url,
-    output: input.output,
-    width: input.viewport.width,
-    height: input.viewport.height,
-    fullPage: input.fullPage,
-    includeImageMetrics: input.includeImageMetrics,
-    includeResourceMetrics: input.includeResourceMetrics,
-    includeContrastMetrics: input.includeContrastMetrics,
-    browser: input.browser ?? "auto",
-    browserPath: input.browserPath,
-    waitUntil: input.waitUntil,
-    waitForSelector: input.waitForSelector,
-    waitForTimeout: input.waitForTimeout,
-    timeout: input.timeout,
-    format: input.format,
-    quality: input.quality,
-    scale: input.scale,
-  });
-  return {
-    async startPreview(input: McpPreviewInput, progress?: McpToolProgress) {
-      await closeCaptureSession();
-      validatePreviewServerOptions({
-        host: input.host ?? "127.0.0.1",
-        port: input.port ?? 5173,
-        imageDomains: input.imageDomains,
-      });
-      progress?.report(
-        "tool preview.start preparing generated preview project"
-      );
-      const previewProject = await preparePreview(
-        input.source,
-        prepareSessionDataFile
-      );
-      progress?.report("tool preview.start starting production preview server");
-      const result = await preview.startAndWait({
-        ...input,
-        cwd: previewProject.cwd,
-        buildCacheKey: previewProject.buildCacheKey,
-        restart: true,
-      });
-      markFresh();
-      return {
-        ...result,
-        generatedBuildMetrics: await inspectGeneratedBuildMetrics(
-          previewProject.cwd
-        ),
-      };
-    },
-    async captureScreenshot(
-      input: McpScreenshotInput,
-      progress?: McpToolProgress
-    ) {
-      const url = await resolveScreenshotUrl(input, progress);
-      progress?.report(`tool screenshot capturing ${url}`);
-      const captureOptions = getCaptureOptions(input, url);
-      if (input.source === "session" && createCaptureSession !== undefined) {
-        captureSession ??= createCaptureSession();
-        return await captureSession.capture(captureOptions);
-      }
-      return await captureScreenshot({
-        ...captureOptions,
-        isJson: false,
-        isMcp: true,
-        isInteractive: false,
-        confirmInstall: async () => false,
-      });
-    },
-    async capturePageScreenshots(
-      inputs: readonly McpScreenshotInput[],
-      progress?: McpToolProgress
-    ) {
-      const firstInput = inputs[0];
-      if (firstInput === undefined) {
-        return [];
-      }
-      if (
-        createCaptureSession === undefined ||
-        firstInput.source !== "session" ||
-        inputs.some((input) => input.source !== firstInput.source)
-      ) {
-        throw new Error(
-          "Resized screenshot capture requires session preview pages."
-        );
-      }
-      const urls: string[] = [];
-      for (const input of inputs) {
-        urls.push(await resolveScreenshotUrl(input, progress));
-      }
-      progress?.report(
-        `tool screenshot capturing ${new Set(urls).size} pages across ${inputs.length} viewport widths`
-      );
-      captureSession ??= createCaptureSession();
-      return await captureSession.capturePage(
-        inputs.map((input, index) => {
-          const url = urls[index];
-          if (url === undefined) {
-            throw new Error("Screenshot URL resolution was incomplete.");
-          }
-          return getCaptureOptions(input, url);
-        })
-      );
-    },
-    async stopPreview() {
-      const errors: unknown[] = [];
-      try {
-        await closeCaptureSession();
-      } catch (error) {
-        errors.push(error);
-      }
-      let result;
-      try {
-        if (preview.stop === undefined) {
-          throw new Error("Preview controller cannot stop its owned preview.");
-        }
-        result = await preview.stop();
-      } catch (error) {
-        errors.push(error);
-      }
-      if (errors.length === 1) {
-        throw errors[0];
-      }
-      if (errors.length > 1) {
-        throw new AggregateError(
-          errors,
-          "Could not clean up screenshot capture or stop the owned preview."
-        );
-      }
-      if (result === undefined) {
-        throw new Error("Preview controller did not return its stopped state.");
-      }
-      return result;
-    },
-  };
 };
 
 export const mcpOptions = (yargs: CommonYargsArgv) =>
@@ -1096,38 +796,6 @@ const assertSingleOpCallToolSupported = (tool: string) => {
   }
 };
 
-export const __testing__ = {
-  createMcpPreviewHandlers,
-  createMcpStatusReporter,
-  formatMcpStatusLine,
-  assertSingleOpCallToolSupported,
-  assertPersistedMcpCheckpointAcknowledged,
-  clearPersistedMcpCheckpoint,
-  createMcpSingleOpCallErrorPayload,
-  createMcpResourceErrorPayload: (error: unknown, elapsedMs: number) => ({
-    ok: false,
-    error: {
-      code: getStableErrorCode(error) ?? "MCP_RESOURCE_FAILED",
-      message: error instanceof Error ? error.message : String(error),
-    },
-    meta: { elapsedMs },
-  }),
-  createMcpRunErrorPayload,
-  createMcpRunCheckpointStopPayload,
-  getLoadedProjectSessionSnapshot,
-  getResultCheckpoint,
-  getMcpOperationInput,
-  parseMcpSingleOpCallInput,
-  validateSingleOpCallInput,
-  isMcpToolCallFailure,
-  getMcpToolCallError,
-  applyMcpRunOptions,
-  parseMcpRunCalls,
-  parseMcpRunInput,
-  readPersistedMcpCheckpoint,
-  updatePersistedMcpCheckpoint,
-};
-
 const createCliMcpHost = async ({
   projectRoot = cwd(),
   projectId,
@@ -1367,7 +1035,37 @@ const executeMcpRunCall = async ({
   return { result, checkpoint };
 };
 
-export const __testingMcpRun = { executeMcpRunCall };
+export const __testing__ = {
+  createMcpStatusReporter,
+  formatMcpStatusLine,
+  assertSingleOpCallToolSupported,
+  assertPersistedMcpCheckpointAcknowledged,
+  clearPersistedMcpCheckpoint,
+  createMcpSingleOpCallErrorPayload,
+  createMcpResourceErrorPayload: (error: unknown, elapsedMs: number) => ({
+    ok: false,
+    error: {
+      code: getStableErrorCode(error) ?? "MCP_RESOURCE_FAILED",
+      message: error instanceof Error ? error.message : String(error),
+    },
+    meta: { elapsedMs },
+  }),
+  createMcpRunErrorPayload,
+  createMcpRunCheckpointStopPayload,
+  getLoadedProjectSessionSnapshot,
+  getResultCheckpoint,
+  getMcpOperationInput,
+  parseMcpSingleOpCallInput,
+  validateSingleOpCallInput,
+  isMcpToolCallFailure,
+  getMcpToolCallError,
+  applyMcpRunOptions,
+  parseMcpRunCalls,
+  parseMcpRunInput,
+  readPersistedMcpCheckpoint,
+  updatePersistedMcpCheckpoint,
+  executeMcpRunCall,
+};
 
 export const mcpSingleOpCall = async (options: McpSingleOpCallOptions) => {
   if (options.tool === undefined || options.tool === "") {
